+++ src/Application/Exception/Participant/ParticipantNotFoundException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Proximum\Vimeet\Application\Exception\Participant;
+
+class ParticipantNotFoundException extends ParticipantException
+{
+}