+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - VIMEET-2117 - Webinar audio/video settings
 - VIMEET-2088 - Sort sheet participants
 - VIMEET-2101 - Add end sound, image and message to visio meeting
+- VIMEET-2105 - Interactive webinar
 
 ## [2.28.0] - 2020-05-18
 ### Added